Add and edit a team result. Remove database .yml from scm.

+++ db/migrate/20150127014554_create_teams.rb
@@ -2,7 +2,7 @@ class CreateTeams < ActiveRecord::Migration
   def change
     create_table :teams do |t|
       t.text :notes
-      t.integer :score
+      t.decimal :score
 
       t.timestamps null: false
     end
